[WEEX-442][Core] Rename interface file.
