+++ weex_core/Source/core/render/action/render_action_add_element.h
@@ -23,7 +23,7 @@
 #include <map>
 #include <set>
 
-#include "core/render/action/render_action.h"
+#include "core/render/action/render_action_interface.h"
 #include "core/layout/style.h"
 
 namespace WeexCore {
