+++ weex_core/Source/core/render/action/render_action_createbody.h
@@ -23,7 +23,7 @@
 #include <map>
 #include <set>
 
-#include "core/render/action/render_action.h"
+#include "core/render/action/render_action_interface.h"
 
 namespace WeexCore {
