+++ weex_core/Source/core/render/action/render_action_layout.h
@@ -22,7 +22,7 @@
 #include <string>
 #include <set>
 
-#include "core/render/action/render_action.h"
+#include "core/render/action/render_action_interface.h"
 
 namespace WeexCore {
