+++ weex_core/Source/core/render/action/render_action_remove_element.h
@@ -21,7 +21,7 @@
 
 #include <string>
 
-#include "core/render/action/render_action.h"
+#include "core/render/action/render_action_interface.h"
 
 namespace WeexCore {
