+++ weex_core/Source/core/render/action/render_action_update_attr.h
@@ -22,7 +22,7 @@
 #include <string>
 #include <vector>
 
-#include "core/render/action/render_action.h"
+#include "core/render/action/render_action_interface.h"
 
 namespace WeexCore {
