+++ weex_core/Source/core/render/node/factory/render_appbar_factory.h
@@ -20,7 +20,7 @@
 #define WEEX_PROJECT_RENDERAPPBARFACTORY_H
 
 #include "core/render/node/render_appbar.h"
-#include "core/render/node/factory/i_render_factory.h"
+#include "core/render/node/factory/render_factory_interface.h"
 
 namespace WeexCore {
