+++ weex_core/Source/core/render/node/factory/render_factory_interface.h
@@ -19,7 +19,7 @@
 #ifndef WEEX_PROJECT_IRENDERFACTORY_H
 #define WEEX_PROJECT_IRENDERFACTORY_H
 
-#include "core/render/node/factory/i_render_object.h"
+#include "core/render/node/factory/render_object_interface.h"
 
 namespace WeexCore {
   class IRenderFactory {
